requirements.txt
python-dotenv
aiogram
gspread
google-auth
pydantic-settings
sqlalchemy
aiosqlite
apscheduler
loguru
